Fixed various bugs, added links interceptor
Related items now show other items which have no relevance when no
relevant items can be found.

Added link interceptor which catches links and shows a warning when the
application is running in a production environment.

Added additions.css, this is a stylesheet which overwrites certain
styles of Khris because they're not compatible with the Multiscreen
Toolkit.